Initial commit
Inside this initial commit I added the Visual Studio 2022 project files whereas the .user files is ignored. Made Core.h to to manage that it is only going to be build for supported platforms (which currently only are Windows). Made an Application class which has a Run() function as a infinite while loop. Made EntryPoint.h to create an basic entry point. Made GitHub issue files which is: Bug Report and Feature Request. Also added LICENSE, README and SECURITY. Fixes: Do something with the main function. It is not critical but it should be done.